Extract BranchDetailsMapper
