+++ Server/Api/Git/Conversions/ServiceRegistration.cs
@@ -1,0 +1,9 @@
+﻿namespace PrincipleStudios.ScaledGitApp.Api.Git.Conversions;
+
+public static class ServiceRegistration
+{
+	internal static void RegisterGitConversions(this IServiceCollection services)
+	{
+		services.AddSingleton<IBranchDetailsMapper, BranchDetailsMapper>();
+	}
+}
