+++ Server/Program.cs
@@ -1,5 +1,6 @@
 #pragma warning disable CA1861 // Avoid constant arrays as arguments - the envfile list belongs here, and static arrays cannot be created in Program.cs
 using dotenv.net;
+using PrincipleStudios.ScaledGitApp.Api.Git.Conversions;
 using PrincipleStudios.ScaledGitApp.Auth;
 using PrincipleStudios.ScaledGitApp.BranchingStrategy;
 using PrincipleStudios.ScaledGitApp.Commands;
@@ -32,6 +33,7 @@ services.RegisterEnvironment(
 	dataProtectionConfig: builder.Configuration.GetSection("DataProtection")
 );
 services.RegisterGit(builder.Configuration.GetSection("git"));
+services.RegisterGitConversions();
 services.RegisterLocales(builder.Configuration.GetSection("localization"));
 services.RegisterRealtimeNotifications(
 	includeAzureSignalR: builder.Configuration["Azure:SignalR:ConnectionString"] != null
